+++ CHANGELOG.md
@@ -10,6 +10,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### v3.9.5 - [Unreleased]
 
+### Changed
+
+- "--letsencrypt=subdomain" option
+
+#### Fixed
+
+- PHP 7.3 extras when php 7.2 isn't installed
+- acme.sh installation
+
 ### v3.9.4 - 2019-03-15
 
 #### Added